Final Fantasy app/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
SQLAlchemy==2.0.36
requests==2.32.3
pywebview==4.4
waitress==3.0.0
pydantic==2.8.2
plotly==5.24.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
reportlab==4.2.3
pytest==8.3.3
kaleido==0.2.1
